requirements.txt
markitdown[pdf]
click
google-generativeai
pytest
